+++ TimeDet/TimeDet.cxx
@@ -23,6 +23,7 @@
 #include "FairRuntimeDb.h"
 #include "FairVolume.h"
 #include "ShipDetectorList.h"
+#include "ShipGeoUtil.h"
 #include "ShipStack.h"
 #include "TClonesArray.h"
 #include "TGeoBBox.h"
@@ -124,26 +125,6 @@ TimeDet::~TimeDet() {
   }
 }
 
-Int_t TimeDet::InitMedium(const char* name) {
-  static FairGeoLoader* geoLoad = FairGeoLoader::Instance();
-  static FairGeoInterface* geoFace = geoLoad->getGeoInterface();
-  static FairGeoMedia* media = geoFace->getMedia();
-  static FairGeoBuilder* geoBuild = geoLoad->getGeoBuilder();
-
-  FairGeoMedium* ShipMedium = media->getMedium(name);
-
-  if (!ShipMedium) {
-    Fatal("InitMedium", "Material %s not defined in media file.", name);
-    return -1111;
-  }
-  TGeoMedium* medium = gGeoManager->GetMedium(name);
-  if (medium != nullptr) return ShipMedium->getMediumIndex();
-
-  return geoBuild->createMedium(ShipMedium);
-
-  return 0;
-}
-
 Bool_t TimeDet::ProcessHits(FairVolume* vol) {
   /** This method is called from the MC stepping */
   // Set parameters at entrance of volume. Reset ELoss.
@@ -219,14 +200,7 @@ void TimeDet::Register() {
 TClonesArray* TimeDet::GetCollection(Int_t iColl) const { return nullptr; }
 
 void TimeDet::UpdatePointTrackIndices(const std::map<Int_t, Int_t>& indexMap) {
-  for (auto& point : *fTimeDetPoints) {
-    Int_t oldTrackID = point.GetTrackID();
-    auto iter = indexMap.find(oldTrackID);
-    if (iter != indexMap.end()) {
-      point.SetTrackID(iter->second);
-      point.SetLink(FairLink("MCTrack", iter->second));
-    }
-  }
+  UpdatePointTrackIndicesImpl(*fTimeDetPoints, indexMap);
 }
 
 void TimeDet::Reset() { fTimeDetPoints->clear(); }
@@ -234,7 +208,7 @@ void TimeDet::Reset() { fTimeDetPoints->clear(); }
 void TimeDet::ConstructGeometry() {
   TGeoVolume* top = gGeoManager->GetTopVolume();
 
-  InitMedium("polyvinyltoluene");
+  ShipGeo::InitMedium("polyvinyltoluene");
   TGeoMedium* Scint = gGeoManager->GetMedium("polyvinyltoluene");
 
   ///////////////////////////////////////////////////////